+++ src/main/java/euteam/cleanmachine/model/billing/AccountSubscription.java
@@ -1,4 +1,4 @@
-package euteam.cleanmachine.model.user;
+package euteam.cleanmachine.model.billing;
 
 import euteam.cleanmachine.model.billing.SubscriptionPlan;
