+++ src/main/java/euteam/cleanmachine/model/billing/SubscriptionPayment.java
@@ -1,8 +1,5 @@
 package euteam.cleanmachine.model.billing;
 
-import euteam.cleanmachine.model.billing.Payment;
-import euteam.cleanmachine.model.user.AccountSubscription;
-
 import javax.persistence.Entity;
 import javax.persistence.ManyToOne;
